+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.8.3 (2018-12-03)
+
+### Updated 
+
+- Support for `symfony/process 4.2`. 
+
 ## 0.8.2 (2018-10-25)
 
 ### Added